pyproject.toml
[tool.black]
    line-length = 99
    target-version = ['py310']

[tool.poetry]
name = "TrainerDex"
version = "4.0.0a7"
description = "An API to interact with TrainerDex - a online database of Pokemon Go trainers"
authors = ["Jay Turner <jay@trainerdex.app>"]
license = "GPL-3.0"
readme = "README.md"
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
homepage = "https://trainerdex.app/"
repository = "https://github.com/TrainerDex/TrainerDex.py"

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = "^3.6.0"
python-dateutil = "^2.8.1"
typing-extensions = "^4.3.0"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
isort = {extras = ["black"], version = "^5.10.1"}
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
